+++ org.dsource.ddt.ide.core.tests/src-test/mmrnmhrm/core/CommonDeeWorkspaceTestNew.java
@@ -13,6 +13,7 @@ package mmrnmhrm.core;
 
 import static melnorme.utilbox.core.Assert.AssertNamespace.assertTrue;
 import melnorme.lang.ide.core.tests.CommonCoreTest;
+import melnorme.lang.ide.core.utils.EclipseUtils;
 import melnorme.utilbox.core.ExceptionAdapter;
 import melnorme.utilbox.misc.StringUtil;
 import mmrnmhrm.tests.CommonDeeWorkspaceTest;
@@ -66,6 +67,10 @@ public abstract class CommonDeeWorkspaceTestNew extends CommonCoreTest {
 	}
 	
 	/* ----------------- ----------------- */
+	
+	public static IProject project(String name) {
+		return EclipseUtils.getProject(name);
+	}
 	
 	public static IProject createLangProject(String name, boolean overwrite) throws CoreException {
 		IProject project = createAndOpenProject(name, overwrite);