+++ org.dsource.ddt.ide.core/src-lang/melnorme/lang/ide/core/utils/EclipseUtils.java
@@ -73,6 +73,7 @@ public class EclipseUtils extends ResourceUtils {
 		project.setDescription(description, null);
 	}
 	
+	@SuppressWarnings("unchecked")
 	public static <T> T getAdapter(Object adaptable, Class<T> adapterType) {
 		return (T) Platform.getAdapterManager().getAdapter(adaptable, adapterType);
 	}